Сomplete3 homework 7.
